requirements.txt
asttokens==2.0.5
backcall==0.2.0
backports.functools-lru-cache==1.6.4
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@b46f5ac7587d2e1862f8b7b1573179d80dcdd620
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
Flask==2.1.2
ftfy==6.1.1
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.11.3
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.0
jupyter-core==4.9.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.5
numpy==1.22.3
packaging
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.8
pyspark==3.0.2
python-dateutil==2.8.2
pyzmq==22.3.0
regex==2022.4.24
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
stack-data==0.2.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.1.2
zipp==3.8.0
